Persist holidays from calendarific and WIP: allow for a reload via UI

+++ src/main/java/za/co/dearx/leave/web/rest/PublicHolidayResource.java
@@ -204,7 +204,7 @@ public class PublicHolidayResource {
     }
 
     @GetMapping("/public-holidays/reload")
-    public ResponseEntity<Void> reloadPublicHolidays(@RequestParam Long yearToLoad) {
+    public ResponseEntity<Void> reloadPublicHolidays(@RequestParam Integer yearToLoad) {
         log.debug("REST request to reload PublicHolidays");
         try {
             publicHolidayService.reloadData(yearToLoad);
